Created static pages for front-end testing purposes
These are the saved HTML from the browser for the pages thus far as rendered by the python server, with a few modifications to turn some buttons into links and replace specific user data with placeholders.

If the repo directory is hosted as an apache site such as on localhost, these will function as a navigable static replica of the live site, for use in front-end testing purposes.

(Hopefully the placement of these files doesn't screw up anything python-wise; I wanted to put them in /static/html/ but then all of their URLs that start from / would be broken, and even if I changed all of those, the CSS etc also makes references to things in /static and so only work properly when the whole site is hosted from the root of the repo, not from somewhere within static, and not with the HTML files just opened from the file system).